% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_safe.R
\name{map_safe_merge}
\alias{map_safe_merge}
\title{Identify location of errors from applying a function to a vector}
\usage{
map_safe_merge(.x, f, ...)
}
\arguments{
\item{.x}{Vector of paths to csvs}

\item{f}{Function that's being applied to the dataset}

\item{...}{Arguments passed to \code{\link{map}}}
}
\value{
A logical vector
}
\description{
Identify location of errors from applying a function to a vector
}
\examples{
if (require(dplyr)){
map_safe_merge(list(starwars, sample_n(mtcars, 1:20), sample_n(mtcars, 1:20)), select, height)
}
}
