% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_safe.R
\name{check_match}
\alias{check_match}
\title{identified whether the user’s requirement existed within the dataset.
if yes, then the function will return true, if not, it returns false.}
\usage{
check_match(.data, ...)
}
\arguments{
\item{.data}{a dataframe}

\item{...}{Arguments}
}
\value{
A logical vector
}
\description{
identified whether the user’s requirement existed within the dataset.
if yes, then the function will return true, if not, it returns false.
}
