% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_safe.R
\name{map_safe}
\alias{map_safe}
\title{Determine if you can apply a function to a vector without errors.}
\usage{
map_safe(.x, f, ...)
}
\arguments{
\item{.x}{Vector of paths to csvs}

\item{f}{Function that's being applied to the dataset}

\item{...}{Arguments}
}
\value{
A logical vector
}
\description{
Determine if you can apply a function to a vector without errors.
}
