% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_safe.R
\name{map_safe_merge}
\alias{map_safe_merge}
\title{Determine if you can apply a function to a vector without errors.}
\usage{
map_safe_merge(.x, f, ...)
}
\arguments{
\item{.x}{Vector of paths to csvs}

\item{f}{Function that's being applied to the dataset}

\item{...}{Arguments passed to \code{\link{map}}}
}
\value{
A logical vector
}
\description{
Determine if you can apply a function to a vector without errors.
}
\examples{
if (require(dplyr)){
map_safe_merge(list(starwars, sample_n(mtcars, 1:20), sample_n(mtcars, 1:20)), select, height)
}
}
